This commit publishes the new dataset.
The new dataset is a correction of the old one, in some sense -- the old dataset was based on incorrect code. The [corrected code](https://github.com/vs-uulm/plexe-veins/tree/vnc2017-cacc-attack-analysis) includes the static attack scenarios (i.e., attacker transmitting a constant value) as originally intended. We've submitted a correction request for the paper.