aggregated datasets
These are generated as a by-product of our plotting scripts, and they
basically provide the input data for our plots. We're publishing these
too, so that others can also have a closer look at the plotted data
found in our paper. These CSVs correspond to the revised dataset (i.e.,
if you run the scripts on the data in `master`, you should end up with
the same CSVs).